Add Eq/PartialEq to API exttypes
Many times it can be quite useful to be able to tell whether two API
exttypes are pointing to the same data or not. So, let's implement
Eq/PartialEq everywhere. We do this using a macro to auto-generate the
trait implementations, as simply deriving them will add an unnecessary
trait bound on <W>.