Always dispatch redraw events sequentially, one at a time
After converting a fork of neovim-gtk that I've been working on over to
using nvim-rs, I noticed some rather strange issues with redrawing. In
particular, initiating completion in cmdline mode would rarely end up
resulting in the status bar being over-written by a line of completions,
followed by a blank line (which should have contained the command being
typed) all the way at the bottom of nvim's screen.

   _________________________________________
  |                                         | <- status line missing
  | foo/ bar/ baz/ biz/ boz/ bogz/ poggers/ |
  | fortnite/ memes/ runningoutofnames/     |
  |_|_______________________________________| <- should be :e

After some investigation, I noticed that nvim-rs dispatches all of its
requests asynchronously. This works fine for most things, with the
particular exception of UI redraw requests. Because each UI redraw request
assumes the state of the screen is equal to the result of the previous
redraw request, the client needs to have the guarantee that the redraw
requests coming in are always coming in the order they were originally sent
out by nvim. Since we currently don't expose the msgid of each notification
to our Handler, not maintaining that order means it's impossible for the
Handler to ascertain the correct order in which to perform each redraw.

So, let's fix this by implementing a queue for redraw requests, which we
use to flush redraw notifications from Neovim sequentially to the client in
parallel with all other RPC notifications.