Serialize requests
Mostly written by KillTheMule, Lyude modified + added tests. This ensures
that we handle notifications in a serialized manner - e.g. we make sure to
only handle one notification at a time, in the order that we received them
in. When handling notifications since the behavior of receiving requests
in the midst of handling a notification is undefined, we currently do not
attempt to handle requests received during the handling of a notification.